Fix earned zero checks in tests
Sometimes the next block was having a slightly higher timestamp, so
the earned reward was not exactly zero. Moved the check to immediately
after the staking transaction, so it’s the same block and therefore
the same time.

Besides, removed `timeIncreaseTo` function. Not sure what it was meant
for, but it was passing a negative value to `setTimeout`, so it seemed buggy.